+++ ImageEvaluator/ImageEvaluatorLib/FindEdgeLines/EdgeLineFinderBase.cs
@@ -24,7 +24,7 @@ namespace ImageEvaluatorLib.FindEdgeLines
         }
 
 
-        public abstract bool Run(Image<Gray, ushort> originalImage, Image<Gray, byte> maskImage, ref IWaferEdgeFindData edgeFindData);
+        public abstract bool Run(Image<Gray, byte> originalImage, Image<Gray, byte> maskImage, ref IWaferEdgeFindData edgeFindData);
 
 
         public virtual bool Init()
@@ -35,7 +35,7 @@ namespace ImageEvaluatorLib.FindEdgeLines
         public bool IsInitialized { get; protected set; }
 
 
-        protected virtual bool CheckInputData(Image<Gray, ushort> originalImage, Image<Gray, byte> maskImage, Dictionary<SearchOrientations, Rectangle> calcAreas)
+        protected virtual bool CheckInputData(Image<Gray, byte> originalImage, Image<Gray, byte> maskImage, Dictionary<SearchOrientations, Rectangle> calcAreas)
         {
             if (originalImage == null || maskImage == null || calcAreas == null)
             {
