+++ ImageEvaluator/ImageEvaluatorLib/ReadDirectory/DirectoryReader.cs
@@ -94,7 +94,7 @@ namespace ImageEvaluatorLib.ReadDirectory
         /// 
         /// </summary>
         /// <returns></returns>
-        public bool GetNextImage(ref Image<Gray, ushort> img1, ref Image<Gray, ushort> img2, ref string name)
+        public bool GetNextImage(ref Image<Gray, byte> img1, ref Image<Gray, byte> img2, ref string name)
         {
             if (!IsInitialized)
             {
