+++ ImageEvaluator/ImageEvaluatorLib/ReadImage/SimpleLightImageReader_Base.cs
@@ -11,7 +11,7 @@ namespace ImageEvaluatorLib.ReadImage
     internal abstract class SimpleLightImageReader_Base : ImageReaderBase
     {
         protected string _fileName;
-        protected Image<Gray, ushort> _img1;
+        protected Image<Gray, byte> _img1;
         protected Rectangle _fullROI;
 
 
@@ -28,7 +28,7 @@ namespace ImageEvaluatorLib.ReadImage
         /// <param name="img1"></param>
         /// <param name="img2">it is not used</param>
         /// <returns></returns>
-        public override bool GetImage(string inputfileName, ref Image<Gray, ushort> img1, ref Image<Gray, ushort> img2)
+        public override bool GetImage(string inputfileName, ref Image<Gray, byte> img1, ref Image<Gray, byte> img2)
         {
 
             if (!IsInitialized)
@@ -115,7 +115,7 @@ namespace ImageEvaluatorLib.ReadImage
 
             try
             {
-                _img1 = new Image<Gray, ushort>(_width, _height);
+                _img1 = new Image<Gray, byte>(_width, _height);
                 _fullROI = new Rectangle(0, 0, _width, _height);
 
                 return true;
